[tools] Update build step to exclude third party
Build systemjs bundle now only includes the d2.angular modules
It tried to also include all the third party modules
after we upgraded the version. The builder now uses its own
config that is defined in the gulpfile. The config.js file
in the root of the project is now only used for the tests.